```text feat: Add Toaster component for displaying toast messages
This commit adds a new component called Toaster, which is used to display toast messages in the application. The Toaster component is imported from the "@/components/ui/toaster" module and is added to the body of the RootLayout component. This feature enhances the user experience by providing a convenient way to show informative messages to the user.