feat: Add useGetCalls hook for fetching calls
The commit adds a new hook called useGetCalls to the project. This hook is responsible for fetching calls from the server using the Stream Video SDK. It utilizes the useUser and useStreamVideoClient hooks from the "@clerk/nextjs" and "@stream-io/video-react-sdk" packages respectively. The hook retrieves calls based on certain filter conditions and sorts them in descending order of their start time. It also categorizes the calls into endedCalls and upcomingCalls based on their start time. This feature enhances the functionality of the application by providing a convenient way to fetch and organize calls.