feat: Add CallList component to Previous page
This commit adds the CallList component to the Previous page in the app/(root)/(home)/previous/page.jsx file. The CallList component is imported from the @components/CallList module and is used to display a list of ended calls. This feature enhances the functionality of the Previous page by showing the user their previous calls in a convenient and organized manner.